# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gemini Chart Analyzer API
  version: 1.0.0
  description: REST API for Gemini Chart Analyzer and Batch Scanner
servers:
  - url: http://localhost:8000/api
paths:
...
  "/analyze/single":
    post:
      tags:
        - Chart Analyzer
      summary: Analyze a single symbol in a single timeframe (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SingleAnalysisRequest"
            example:
              symbol: "BTC/USDT"
              timeframe: "1h"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeRequestResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  "/analyze/multi":
    post:
      tags:
        - Chart Analyzer
      summary: Analyze a symbol across multiple timeframes (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiAnalysisRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeRequestResponse"
  "/analyze/{session_id}/status":
    get:
      tags:
        - Chart Analyzer
      summary: Get analysis status
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeStatusResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
...
    IndicatorsConfig:
      type: object
      properties:
        ma_periods:
          type: array
          items:
            type: integer
        rsi_period:
          type: integer
        enable_macd:
          type: boolean
        enable_bb:
          type: boolean
        bb_period:
          type: integer
    SingleAnalysisRequest:
      type: object
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        indicators:
          $ref: "#/components/schemas/IndicatorsConfig"
        prompt_type:
          type: string
        custom_prompt:
          type: string
        limit:
          type: integer
        chart_figsize:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
        chart_dpi:
          type: integer
        no_cleanup:
          type: boolean
      required:
        - symbol
        - timeframe
    MultiAnalysisRequest:
      type: object
      properties:
        symbol:
          type: string
        timeframes:
          type: array
          items:
            type: string
        indicators:
          $ref: "#/components/schemas/IndicatorsConfig"
        prompt_type:
          type: string
        custom_prompt:
          type: string
        limit:
          type: integer
        chart_figsize:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
        chart_dpi:
          type: integer
        no_cleanup:
          type: boolean
      required:
        - symbol
        - timeframes
...
    AnalyzeRequestResponse:
      type: object
      properties:
        success:
          type: boolean
        session_id:
          type: string
        status:
          type: string
        message:
          type: string
    AnalyzeStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        session_id:
          type: string
        status:
          type: string
        started_at:
          type: string
        completed_at:
          type: string
        result:
          type: object
        error:
          type: string
...