# web/api/chart_analyzer.py
"""
API routes for Chart Analyzer (single and multi-timeframe analysis).
"""

import gc
import os
import uuid
import time
import re
import json
from pathlib import Path
from typing import Optional, Dict, List, Tuple
from urllib.parse import quote
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel, Field

from modules.common.core.exchange_manager import ExchangeManager
from modules.common.core.data_fetcher import DataFetcher
from modules.gemini_chart_analyzer.core.generators.chart_generator import ChartGenerator
from modules.gemini_chart_analyzer.core.analyzers.gemini_chart_analyzer import GeminiChartAnalyzer
from modules.gemini_chart_analyzer.core.analyzers.multi_timeframe_coordinator import MultiTimeframeCoordinator
from modules.gemini_chart_analyzer.core.utils.chart_paths import get_charts_dir
from modules.common.utils import normalize_timeframe
from modules.common.ui.logging import log_error
from web.utils.log_manager import get_log_manager
from web.utils.cli_logger import CLILogger
from web.utils.task_manager import get_task_manager

router = APIRouter()


# Request/Response models
class IndicatorsConfig(BaseModel):
    """Indicators configuration."""
    ma_periods: Optional[List[int]] = Field(default=[20, 50, 200], description="Moving Average periods")
    rsi_period: Optional[int] = Field(default=14, description="RSI period")
    enable_macd: Optional[bool] = Field(default=True, description="Enable MACD")
    enable_bb: Optional[bool] = Field(default=False, description="Enable Bollinger Bands")
    bb_period: Optional[int] = Field(default=20, description="Bollinger Bands period")


class SingleAnalysisRequest(BaseModel):
    """Request model for single timeframe analysis."""
    symbol: str = Field(..., description="Trading symbol (e.g., BTC/USDT)")
    timeframe: str = Field(..., description="Timeframe (e.g., 1h, 4h, 1d)")
    indicators: Optional[IndicatorsConfig] = Field(default_factory=IndicatorsConfig)
    prompt_type: Optional[str] = Field(default="detailed", description="Prompt type: detailed, simple, or custom")
    custom_prompt: Optional[str] = Field(default=None, description="Custom prompt (if prompt_type is custom)")
    limit: Optional[int] = Field(default=500, description="Number of candles to fetch")
    chart_figsize: Optional[Tuple[int, int]] = Field(default=(16, 10), description="Chart figure size (width, height)")
    chart_dpi: Optional[int] = Field(default=150, description="Chart DPI")
    no_cleanup: Optional[bool] = Field(default=False, description="Don't cleanup old charts")


class MultiAnalysisRequest(BaseModel):
    """Request model for multi-timeframe analysis."""
    symbol: str = Field(..., description="Trading symbol (e.g., BTC/USDT)")
    timeframes: List[str] = Field(..., description="List of timeframes (e.g., ['15m', '1h', '4h', '1d'])")
    indicators: Optional[IndicatorsConfig] = Field(default_factory=IndicatorsConfig)
    prompt_type: Optional[str] = Field(default="detailed", description="Prompt type: detailed, simple, or custom")
    custom_prompt: Optional[str] = Field(default=None, description="Custom prompt (if prompt_type is custom)")
    limit: Optional[int] = Field(default=500, description="Number of candles to fetch")
    chart_figsize: Optional[Tuple[int, int]] = Field(default=(16, 10), description="Chart figure size (width, height)")
    chart_dpi: Optional[int] = Field(default=150, description="Chart DPI")
    no_cleanup: Optional[bool] = Field(default=False, description="Don't cleanup old charts")


def _convert_indicators_config(config: IndicatorsConfig) -> Dict:
    """Convert IndicatorsConfig to indicators dict format."""
    indicators = {}
    
    if config.ma_periods:
        indicators['MA'] = {'periods': config.ma_periods}
    
    if config.rsi_period:
        indicators['RSI'] = {'period': config.rsi_period}
    
    if config.enable_macd:
        indicators['MACD'] = {'fast': 12, 'slow': 26, 'signal': 9}
    
    if config.enable_bb:
        indicators['BB'] = {'period': config.bb_period or 20, 'std': 2}
    
    return indicators


def _cleanup_old_charts(charts_dir: Path, max_age_seconds: int = 3600):
    """
    Cleanup old chart files.

    Only deletes charts older than `max_age_seconds` (default 1 hour).
    This reduces risk of interfering with currently-active charts in a multi-user environment.
    """
    now = time.time()
    try:
        if charts_dir.exists():
            for file in charts_dir.glob("*.png"):
                try:
                    # Remove only files older than max_age_seconds
                    file_stat = file.stat()
                    file_age = now - file_stat.st_mtime
                    if file_age > max_age_seconds:
                        file.unlink()
                except Exception:
                    pass
    except Exception:
        pass


def _get_chart_url(chart_path: str) -> Optional[str]:
    """Convert chart path to URL."""
    if not chart_path:
        return None
    
    # Get relative path from charts directory
    charts_dir = get_charts_dir()
    try:
        rel_path = os.path.relpath(chart_path, str(charts_dir))
        # Normalize path separators
        rel_path = rel_path.replace('\\', '/')
        # URL encode
        parts = rel_path.split('/')
        encoded_parts = [quote(part) for part in parts]
        return f"/static/charts/{'/'.join(encoded_parts)}"
    except Exception:
        # Fallback: just use filename
        filename = os.path.basename(chart_path)
        return f"/static/charts/{filename}"


@router.post("/analyze/single")
async def analyze_single(request: SingleAnalysisRequest):
    """
    Analyze a single symbol on a single timeframe.
    
    Runs in background thread and returns session_id immediately.
    
    Returns:
        Session ID and status. Use GET /api/analyze/{session_id}/status to check progress.
    """
    try:
        # Generate session_id immediately
        session_id = str(uuid.uuid4())
        
        # Normalize timeframe
        timeframe = normalize_timeframe(request.timeframe)
        
        # Create log file
        log_manager = get_log_manager()
        log_manager.create_log_file(session_id, "analyze")
        
        # Get task manager
        task_manager = get_task_manager()
        
        # Define background task function
        def run_analysis():
            """Run analysis in background thread with logging."""
            cli_logger = CLILogger(session_id, "analyze")
            
            try:
                with cli_logger.capture_output():
                    # Initialize components
                    exchange_manager = ExchangeManager()
                    data_fetcher = DataFetcher(exchange_manager)
                    
                    # Fetch OHLCV data
                    df, exchange_id = data_fetcher.fetch_ohlcv_with_fallback_exchange(
                        symbol=request.symbol,
                        timeframe=timeframe,
                        limit=request.limit,
                        check_freshness=False
                    )
                    
                    if df is None or df.empty:
                        raise ValueError(f"Không thể lấy dữ liệu OHLCV cho {request.symbol} ({timeframe})")
                    
                    # Cleanup old charts if needed
                    if not request.no_cleanup:
                        charts_dir = get_charts_dir()
                        _cleanup_old_charts(charts_dir)
                    
                    # Convert indicators config
                    indicators = _convert_indicators_config(request.indicators)
                    
                    # Generate chart
                    chart_generator = ChartGenerator(
                        figsize=request.chart_figsize,
                        style='dark_background',
                        dpi=request.chart_dpi
                    )
                    
                    chart_path = chart_generator.create_chart(
                        df=df,
                        symbol=request.symbol,
                        timeframe=timeframe,
                        indicators=indicators or None,
                        show_volume=True,
                        show_grid=True
                    )
                    
                    # Analyze with Gemini
                    gemini_analyzer = GeminiChartAnalyzer()
                    analysis_result = gemini_analyzer.analyze_chart(
                        image_path=chart_path,
                        symbol=request.symbol,
                        timeframe=timeframe,
                        prompt_type=request.prompt_type,
                        custom_prompt=request.custom_prompt
                    )
                    
                    # Extract signal and confidence from analysis
                    signal = "NONE"
                    confidence = 0.0

                    # Convert analysis_result to string and lowercase for pattern matching
                    # Convert to string if needed
                    if isinstance(analysis_result, str):
                        analysis_lower = analysis_result.lower()
                    else:
                        analysis_lower = str(analysis_result).lower()

                    # Use regex and negation keywords for better signal extraction
                    def contains_negation(context, keyword_idx, window=8):
                        # Look for 'not', 'no', 'none', 'avoid', 'without' within N words before keyword
                        neg_words = r'(not|no|none|avoid|without|never|neither|fail(ed)? to)'
                        # Find substring window before the keyword
                        start = max(0, keyword_idx - window)
                        pre_context = context[start:keyword_idx]
                        result = re.search(rf'\b{neg_words}\b', pre_context)
                        return result

                    patterns = [
                        # Pattern for positive long, not negated
                        (r'\blong\b', "LONG"),
                        (r'\bshort\b', "SHORT")
                    ]

                    for pat, candidate_signal in patterns:
                        match = re.search(pat, analysis_lower)
                        if match:
                            idx = match.start()
                            # If there's a negation before the keyword, skip
                            if contains_negation(analysis_lower, idx):
                                continue
                            
                            # Also check for phrases like "not a long signal", "avoid short position"
                            # If negation found after pattern, skip as well
                            end_window = 15
                            end = match.end()
                            post_context = analysis_lower[end:end+end_window]
                            if re.search(r'\bnot\b|\bavoid\b|\bno\b', post_context):
                                continue
                            
                            # No negation found, extract signal
                            signal = candidate_signal
                            confidence = 0.7
                            break

                    
                    # Convert chart path to URL
                    chart_url = _get_chart_url(chart_path)
                    
                    # Prepare response
                    response = {
                        "success": True,
                        "symbol": request.symbol,
                        "timeframe": timeframe,
                        "analysis": analysis_result,
                        "chart_path": chart_path,
                        "chart_url": chart_url,
                        "signal": signal,
                        "confidence": confidence,
                        "exchange": exchange_id
                    }
                    
                    # Save result to task manager
                    task_manager.set_result(session_id, response)
                    
            except Exception as e:
                error_msg = str(e)
                log_error(f"Analysis error for session {session_id}: {error_msg}")
                task_manager.set_error(session_id, error_msg)
            finally:
                # Cleanup resources to free memory
                try:
                    gc.collect()
                except Exception:
                    pass
        
        # Start background task
        task_manager.start_task(session_id, run_analysis, "analyze")
        
        # Return immediately with session_id
        return {
            "success": True,
            "session_id": session_id,
            "status": "running",
            "message": "Analysis started. Use GET /api/analyze/{session_id}/status to check progress."
        }
        
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Lỗi khi khởi động phân tích: {str(e)}")


@router.post("/analyze/multi")
async def analyze_multi(request: MultiAnalysisRequest):
    """
    Analyze a single symbol across multiple timeframes.
    
    Runs in background thread and returns session_id immediately.
    
    Returns:
        Session ID and status. Use GET /api/analyze/{session_id}/status to check progress.
    """
    try:
        # Generate session_id immediately
        session_id = str(uuid.uuid4())
        
        # Normalize timeframes
        from modules.gemini_chart_analyzer.core.utils import normalize_timeframes, validate_timeframes
        
        is_valid, error_msg = validate_timeframes(request.timeframes)
        if not is_valid:
            raise HTTPException(status_code=400, detail=error_msg or "Invalid timeframes")
        
        timeframes_list = normalize_timeframes(request.timeframes)
        
        # Create log file
        log_manager = get_log_manager()
        log_manager.create_log_file(session_id, "analyze")
        
        # Get task manager
        task_manager = get_task_manager()
        
        # Define background task function
        def run_analysis():
            """Run multi-timeframe analysis in background thread with logging."""
            cli_logger = CLILogger(session_id, "analyze")
            
            try:
                with cli_logger.capture_output():
                    # Initialize components
                    exchange_manager = ExchangeManager()
                    data_fetcher = DataFetcher(exchange_manager)
                    
                    # Cleanup old charts if needed
                    if not request.no_cleanup:
                        charts_dir = get_charts_dir()
                        _cleanup_old_charts(charts_dir)
                    
                    # Convert indicators config
                    indicators = _convert_indicators_config(request.indicators)
                    
                    # Initialize multi-timeframe analyzer
                    mtf_analyzer = MultiTimeframeCoordinator()
                    
                    # Define helper functions
                    def fetch_data_func(sym, tf):
                        df, _ = data_fetcher.fetch_ohlcv_with_fallback_exchange(
                            symbol=sym,
                            timeframe=tf,
                            limit=request.limit,
                            check_freshness=False
                        )
                        return df
                    
                    def generate_chart_func(df, sym, tf):
                        chart_gen = ChartGenerator(
                            figsize=request.chart_figsize,
                            style='dark_background',
                            dpi=request.chart_dpi
                        )
                        return chart_gen.create_chart(
                            df=df,
                            symbol=sym,
                            timeframe=tf,
                            indicators=indicators or None,
                            show_volume=True,
                            show_grid=True
                        )
                    
                    # Create analyzer once for reuse
                    gemini_analyzer = GeminiChartAnalyzer()                    
                    def analyze_chart_func(chart_path, sym, tf):
                        return gemini_analyzer.analyze_chart(
                            image_path=chart_path,
                            symbol=sym,
                            timeframe=tf,
                            prompt_type=request.prompt_type,
                            custom_prompt=request.custom_prompt
                        )
                    
                    # Run multi-timeframe analysis
                    results = mtf_analyzer.analyze_deep(
                        symbol=request.symbol,
                        timeframes=timeframes_list,
                        fetch_data_func=fetch_data_func,
                        generate_chart_func=generate_chart_func,
                        analyze_chart_func=analyze_chart_func
                    )
                    
                    # Convert chart paths to URLs
                    timeframe_results = {}
                    for tf in timeframes_list:
                        if tf in results.get('timeframes', {}):
                            tf_result = results['timeframes'][tf].copy()
                            chart_path = tf_result.get('chart_path')
                            if chart_path:
                                tf_result['chart_url'] = _get_chart_url(chart_path)
                            timeframe_results[tf] = tf_result
                    
                    # Get aggregated results
                    aggregated = results.get('aggregated', {})
                    
                    # Prepare response
                    response = {
                        "success": True,
                        "symbol": request.symbol,
                        "timeframes": timeframes_list,
                        "timeframes_results": timeframe_results,
                        "aggregated": {
                            "signal": aggregated.get('signal', 'NONE'),
                            "confidence": aggregated.get('confidence', 0.0),
                            "weights_used": aggregated.get('weights_used', {})
                        },
                        "full_results": results
                    }
                    
                    # Save result to task manager
                    task_manager.set_result(session_id, response)
                    
            except Exception as e:
                error_msg = str(e)
                log_error(f"Multi-timeframe analysis error for session {session_id}: {error_msg}")
                task_manager.set_error(session_id, error_msg)
            finally:
                # Cleanup resources to free memory
                try:
                    import gc
                    gc.collect()
                except Exception:
                    pass
        
        # Start background task
        task_manager.start_task(session_id, run_analysis, "analyze")
        
        # Return immediately with session_id
        return {
            "success": True,
            "session_id": session_id,
            "status": "running",
            "message": "Multi-timeframe analysis started. Use GET /api/analyze/{session_id}/status to check progress."
        }
        
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Lỗi khi khởi động phân tích: {str(e)}")


@router.get("/analyze/{session_id}/status")
async def get_analyze_status(session_id: str):
    """
    Get status of an analysis task.
    
    Args:
        session_id: Session ID from analyze_single or analyze_multi endpoint
        
    Returns:
        Status and results (if completed)
    """
    try:
        task_manager = get_task_manager()
        status = task_manager.get_status(session_id)
        
        # Check if session doesn't exist
        if status is None:
            raise HTTPException(status_code=404, detail=f"Session {session_id} not found")
        
        # Extract status information
        status_value = status.get('status')
        started_at = status.get('started_at')
        completed_at = status.get('completed_at')
        
        # Build response
        response = {
            "success": True,
            "session_id": str(session_id),
            "status": status_value,
            "started_at": started_at.isoformat() if started_at is not None else None,
            "completed_at": completed_at.isoformat() if completed_at is not None else None,
        }
        
        # Include result if completed
        if status_value == 'completed' and status.get('result'):
            try:
                # Convert to JSON-serializable format
                serialized = json.dumps(status['result'], default=str, ensure_ascii=False)
                response['result'] = json.loads(serialized)
            except (TypeError, ValueError, RecursionError):
                response['result'] = str(status['result']) if status['result'] else None
        
        # Include error if failed
        if status_value == 'error' and status.get('error'):
            try:
                serialized = json.dumps(status['error'], default=str, ensure_ascii=False)
                response['error'] = json.loads(serialized)
            except (TypeError, ValueError, RecursionError):
                response['error'] = str(status['error']) if status['error'] else None
        
        return response
        
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Lỗi khi lấy status: {str(e)}")


@router.post("/analyze/{session_id}/cancel")
async def cancel_analysis(session_id: str):
    """
    Cancel a running analysis task.
    
    Args:
        session_id: Session ID from analyze_single or analyze_multi endpoint
        
    Returns:
        Success message and cancelled status
    """
    try:
        task_manager = get_task_manager()
        
        # Check if task exists
        status = task_manager.get_status(session_id)
        if status is None:
            raise HTTPException(
                status_code=404,
                detail=f"Session {session_id} not found"
            )
        
        # Check if task can be cancelled
        current_status = status.get('status')
        if current_status not in ['running']:
            raise HTTPException(
                status_code=400,
                detail=f"Cannot cancel task with status '{current_status}'. Only running tasks can be cancelled."
            )
        
        # Cancel task
        cancelled = task_manager.cancel_task(session_id)
        if not cancelled:
            raise HTTPException(
                status_code=400,
                detail=f"Failed to cancel task {session_id}"
            )
        
        return {
            "success": True,
            "session_id": session_id,
            "status": "cancelled",
            "message": "Analysis cancelled successfully. The task will stop processing after current operation."
        }
        
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Lỗi khi cancel analysis: {str(e)}")